requirements.txt
fastapi
uvicorn
openai
google-genai
python-dotenv
starlette
requests
Pillow
boto3
aiohttp